Fix all build errors across services and models
- Exclude Tests/ from main project compilation to prevent missing xunit/FluentAssertions references
- Rename ValidationMemoryService methods to match interface (GetSessionPhaseStatusAsync, UpdateSessionPhaseStatusAsync)
- Fix MetricsService: correct SchemaAnalysisResult, TableComparisonResult, DiscrepancyAnalysisResult, and MigrationRiskAssessment property references
- Fix DiscrepancyDetector: cast long→int for DetermineSeverity calls, fix DifferenceSeverity enum switch, fix SchemaCompatibility type references, fix SampleDifferences usage
- Fix DashboardService/ProgressService: replace untyped Task.FromResult(new()) with return new()
- Fix RiskCalculator: rewrite non-constant switch expression as if-else chain
- Fix StatusCacheService: fix generic new() constraint error, map Revalidator→WarmupFunction
- Fix ReportGenerator: cast long→int for ComparisonSummaryReport fields
- Add missing model properties: DiscoveryFindings (TablesAnalyzed, ColumnsAnalyzed, etc.), ExecutionResults (StartTime, EndTime, DurationMs), ExecutiveSummary (GoBlockers), CacheStatistics (TotalGetRequests, TotalSetRequests, TotalDeleteRequests, long types for hits/misses/evictions)

+++ ValidationOrchestrator/Comparison/ReportGenerator.cs
@@ -199,11 +199,11 @@ public class ReportGenerator : IReportGenerator
             TablesMatched = comparisons?.Count(c => c.Status == ComparisonStatus.Identical) ?? 0,
             TablesWithDifferences = comparisons?.Count(c => c.Status != ComparisonStatus.Identical) ?? 0,
             OverallMatchPercentage = comparisons?.Average(c => c.MatchPercentage) ?? 0,
-            TotalRecordsCompared = comparisons?.Sum(c => c.LegacyRecordCount) ?? 0,
-            MatchedRecords = comparisons?.Sum(c => c.MatchedRecords) ?? 0,
-            MissingRecords = comparisons?.Sum(c => c.MissingRecords) ?? 0,
-            ExtraRecords = comparisons?.Sum(c => c.ExtraRecords) ?? 0,
-            ModifiedRecords = comparisons?.Sum(c => c.ModifiedRecords) ?? 0
+            TotalRecordsCompared = (int)(comparisons?.Sum(c => c.LegacyRecordCount) ?? 0),
+            MatchedRecords = (int)(comparisons?.Sum(c => c.MatchedRecords) ?? 0),
+            MissingRecords = (int)(comparisons?.Sum(c => c.MissingRecords) ?? 0),
+            ExtraRecords = (int)(comparisons?.Sum(c => c.ExtraRecords) ?? 0),
+            ModifiedRecords = (int)(comparisons?.Sum(c => c.ModifiedRecords) ?? 0)
         };
     }
